Add security provider earlier in bootstrap process (#5749)
(cherry picked from commit 6f2b39a6bfd826622289afd4c3728adcc4bcfa49)
