+++ src/main/java/org/opensearch/security/ssl/OpenSearchSecuritySSLPlugin.java
@@ -20,6 +20,7 @@ package org.opensearch.security.ssl;
 import java.nio.file.Path;
 import java.security.AccessController;
 import java.security.PrivilegedAction;
+import java.security.Security;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -34,6 +35,7 @@ import java.util.function.Supplier;
 import com.fasterxml.jackson.databind.InjectableValues;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.bouncycastle.jcajce.provider.BouncyCastleFipsProvider;
 
 import org.opensearch.OpenSearchException;
 import org.opensearch.SpecialPermission;
@@ -255,6 +257,8 @@ public class OpenSearchSecuritySSLPlugin extends Plugin implements SystemIndexPl
         if (!httpSSLEnabled && !transportSSLEnabled) {
             log.error("SSL not activated for http and/or transport.");
         }
+
+        tryAddSecurityProvider();
 
         this.sslSettingsManager = new SslSettingsManager(new Environment(settings, configPath));
     }
@@ -771,5 +775,15 @@ public class OpenSearchSecuritySSLPlugin extends Plugin implements SystemIndexPl
 
     public ThreadPool getThreadPool() {
         return this.threadPool;
+    }
+
+    private void tryAddSecurityProvider() {
+        AccessController.doPrivileged(() -> {
+            if (Security.getProvider("BCFIPS") == null) {
+                Security.addProvider(new BouncyCastleFipsProvider());
+                log.debug("Bouncy Castle FIPS Provider added");
+            }
+            return null;
+        });
     }
 }
